+++ src/store/index.js
@@ -6,7 +6,9 @@ Vue.use(Vuex)
 export default new Vuex.Store({
   state: {
     drawerState: true,
-    basemapState: "Topo"
+    basemapState: "Topo",
+    streamgageState: false,
+    currentZoomState: 4
   },
   mutations: {
     toggleDrawerState (state, data) {
@@ -14,11 +16,19 @@ export default new Vuex.Store({
     },
     getBasemapState (state, selected){
       state.basemapState = selected
+    },
+    getStreamgageState (state, streamgagePicked){
+      state.streamgageState = streamgagePicked
+    },
+    getCurrentZoomState (state, currentZoom){
+      state.currentZoomState = currentZoom
     }
   },
   getters : {
     drawerState: (state) => state.drawerState,
-    basemapState: (state) => state.basemapState
+    basemapState: (state) => state.basemapState,
+    streamgageState: (state) => state.streamgageState,
+    currentZoomState: (state) => state.currentZoomState
   },
   actions: {
   },